fix(file): rename go1.2 platform dir to go1.26
The ballerina/file stdlib platform directory was mistakenly named
go1.2, diverging from every other stdlib's go1.26 convention. Rename
the directory and fix the stray go1.2 references in the native import
path, Bala.toml platform field, and test phase table.
