+++ lib/rt/libs.go
@@ -33,7 +33,7 @@ import (
 	// standard libraries
 	_ "github.com/ballerina-nutcracker/ballerina/lib/stdlibs/ballerina/avro/0.0.1/go1.26/native"
 	_ "github.com/ballerina-nutcracker/ballerina/lib/stdlibs/ballerina/crypto/0.0.1/go1.26/native"
-	_ "github.com/ballerina-nutcracker/ballerina/lib/stdlibs/ballerina/file/0.0.1/go1.2/native"
+	_ "github.com/ballerina-nutcracker/ballerina/lib/stdlibs/ballerina/file/0.0.1/go1.26/native"
 	_ "github.com/ballerina-nutcracker/ballerina/lib/stdlibs/ballerina/http/0.0.1/go1.26/native"
 	_ "github.com/ballerina-nutcracker/ballerina/lib/stdlibs/ballerina/io/0.0.1/go1.26/native"
 	_ "github.com/ballerina-nutcracker/ballerina/lib/stdlibs/ballerina/log/0.0.1/go1.26/native"
