+++ lib/stdlibs/ballerina/file/0.0.1/go1.26/Bala.toml
@@ -5,7 +5,7 @@ schema_version = "4"
 ballerina_version      = ""
 implementation_vendor  = "WSO2"
 language_spec_version  = "2024R1"
-platform               = "go1.2"
+platform               = "go1.26"
 
 [[modules]]
 name   = "file"
